Fix network switch bug for generic EIP-1193 providers
A EIP-1193 spec does not describe an "isConnected()" method,
so if this method is not found on the provider,
the dapp should not assume it is not connected. Instead, only perform
the check if the method is found.
This fixed a bug when the UI goes into a infinite cycle of switching
the network back and forth.